+++ src/mailer.ts
@@ -1,7 +1,10 @@
-import { render } from 'ejs';
-import { readFileSync } from 'fs';
 import nodemailer from 'nodemailer';
 import Mail from 'nodemailer/lib/mailer';
+import { join } from 'path';
+import {
+  parseCompositeTemplate,
+  parseTemplatesDirectly
+} from './template-utils';
 import { Config } from './types/config';
 
 export class Mailer {
@@ -21,63 +24,51 @@ export class Mailer {
   }
 
   /**
-   * Use the same nodemailer config that Superlogin uses to send out an email.
-   * @param templateName the entry under the `emails` property in the config
+   * Use the same nodemailer config that couch-auth uses to send out an email.
+   * Internally, `req` (the sent request), `user` (the sl-user doc) and `data`
+   * defined in the config are always available.
+   *
+   * @param templateId the entry under the `emails` property in the config
    * @param recepient the recepient's email address
-   * @param data additional data can be passed to `ejs.render()`
+   * @param data additional data can be passed to `nunjucks.render()`
    */
-  sendEmail(
-    templateName: string,
-    recepient: string,
-    data?: Record<string, any>
-  ) {
+  sendEmail(templateId: string, recepient: string, data?: Record<string, any>) {
     // load the template and parse it
-    let templateFiles = this.config.emails[templateName]?.templates;
-    if (!templateFiles) {
-      const templateFile = this.config.emails[templateName]?.template;
-      if (!templateFile) {
-        return Promise.reject('No templates found for "' + templateName + '".');
-      }
-      templateFiles = [templateFile];
+    const templateConfig = this.config.emailTemplates[templateId];
+    if (!templateConfig) {
+      return Promise.reject('No template entry for "' + templateId + '".');
     }
+    const templateDirectory =
+      this.config.emailTemplateFolder ?? join(__dirname, './templates/email');
+    let templates: { html: string; text: string };
+    const templateData = { ...templateConfig, ...data };
 
-    const readTemplates = templateFiles.map(t => readFileSync(t, 'utf8'));
-    for (let i = 0; i < templateFiles.length; i++) {
-      if (!readTemplates[i]) {
-        return Promise.reject(
-          'Failed to locate template file: ' + templateFiles[i]
-        );
+    try {
+      templates = parseCompositeTemplate(
+        templateDirectory,
+        templateId,
+        templateData
+      );
+    } catch (error) {
+      templates = parseTemplatesDirectly(
+        templateDirectory,
+        templateId,
+        templateData
+      );
+      if (!templates.html && !templates.text) {
+        return Promise.reject(`No template file found for "${templateId}"`);
       }
     }
-    const renderedTemplates = readTemplates.map(t => render(t, data));
 
     // form the email
-    const subject = this.config.emails[templateName].subject;
-    let formats = this.config.emails[templateName].formats;
-    if (!formats) {
-      const format = this.config.emails[templateName].format;
-      if (!format) {
-        return Promise.reject('No formats specified for: ' + templateName);
-      }
-      formats = [format];
-    }
-    if (formats.length !== renderedTemplates.length) {
-      return Promise.reject(
-        'Different number of read templates and requested formats for template: ' +
-          templateName
-      );
-    }
     let mailOptions: Mail.Options = {
       from: this.config.mailer.fromEmail,
       to: recepient,
-      subject: subject
+      subject: templateConfig.subject,
+      ...templates
     };
     if (this.config.mailer.messageConfig) {
       mailOptions = { ...this.config.mailer.messageConfig, ...mailOptions };
-    }
-
-    for (let i = 0; i < formats.length; i++) {
-      mailOptions[formats[i]] = renderedTemplates[i];
     }
     if (this.config.testMode?.debugEmail) {
       console.log(mailOptions);