+++ test/mailer.spec.ts
@@ -1,8 +1,8 @@
 'use strict';
-import { ConfigHelper as Configure } from '../src/config/configure';
 import { expect } from 'chai';
+import { join } from 'path';
+import { ConfigHelper as Configure } from '../src/config/configure';
 import { Mailer } from '../src/mailer';
-import path from 'path';
 
 const mailerTestConfig = new Configure({
   testMode: {
@@ -11,13 +11,7 @@ const mailerTestConfig = new Configure({
   mailer: {
     fromEmail: 'noreply@example.com'
   },
-  emails: {
-    confirmEmail: {
-      subject: 'Please confirm your email',
-      template: path.join(__dirname, '../templates/email/confirm-email.ejs'),
-      format: 'text'
-    }
-  }
+  emailTemplateFolder: join(__dirname, '../templates/email')
 });
 
 const req = {
@@ -27,7 +21,7 @@ const req = {
   }
 };
 
-const theUser = {
+const user = {
   name: 'Super',
   unverifiedEmail: {
     token: 'abc123'
@@ -40,8 +34,8 @@ describe('Mailer', function () {
   it('should send a confirmation email', function () {
     return mailer
       .sendEmail('confirmEmail', 'super@example.com', {
-        req: req,
-        user: theUser
+        req,
+        user
       })
       .then(function (result) {
         const response = result.response.toString();
@@ -57,5 +51,20 @@ describe('Mailer', function () {
           response.search('https://example.com/auth/confirm-email/abc123')
         ).to.be.greaterThan(-1);
       });
+  });
+
+  it('should render all default templates', () => {
+    const data = {
+      req,
+      user
+    };
+    let templateCount = 0;
+    for (const template of Object.keys(
+      mailerTestConfig.config.emailTemplates
+    )) {
+      mailer.sendEmail(template, 'super@example.com', data);
+      templateCount += 1;
+    }
+    expect(templateCount).to.be.equal(6);
   });
 });